refactor: tests scripts structure
