+++ tests/run_knowledge_base_sheets_tests.py
@@ -1,9 +1,9 @@
 import logging
 from datetime import datetime
 
-from api_service_kb import authenticate, start_knowledge_base_prediction
-from google_sheets_service_kb import get_test_cases, create_new_sheet_and_write_results
-from result_knowledge_base import process_results
+from services.api_service_kb import authenticate, start_knowledge_base_prediction
+from services.google_sheets_service_kb import get_test_cases, create_new_sheet_and_write_results
+from utils.result_knowledge_base import process_results
 
 # Настройка логгирования
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
